pricing-engine/requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
shap
redis
psycopg2-binary
pandas
numpy
requests
python-dotenv